# [isl] ci: fix linter issues related to dismissBugDropdown